flower-search/requirements.txt
click
pillow
jina[devel]==0.4.1
